+++ CustomizeExterior/Framework/SavedExteriors.cs
@@ -1,9 +1,0 @@
-using System.Collections.Generic;
-
-namespace CustomizeExterior.Framework
-{
-    internal class SavedExteriors
-    {
-        public Dictionary<string, string> Chosen = new();
-    }
-}